pyproject.toml
[tool.poetry]
name = "deep-learning"
version = "0.1.0"
description = ""
authors = ["errolelliott <ezzy.elliott@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.2"
h5py = "^3.12.1"
matplotlib = "^3.9.2"
pillow = "^10.4.0"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

